refactor: 각 도메인별로 Fixtures를 분리하여 관리
