+++ src/test/java/kitchenpos/fixtures/ProductFixtures.java
@@ -34,4 +34,13 @@ public class ProductFixtures {
 
         return product;
     }
+
+    public static Product createProduct(final String name, final BigDecimal price) {
+        Product product = new Product();
+        product.setId(UUID.randomUUID());
+        product.setName(name);
+        product.setPrice(price);
+        return product;
+    }
+
 }
